ai_services/ai_chatbot/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = []

[project]
name = "ai-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "openai-agents>=0.9.0",
    "pinecone>=8.0.1",
    "python-docs>=0.1.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
]
